Improve the QE comment integration
Handles the case of someone opening the QE for one email, closing it, changing the email, and then trying to open the QE again.
