+++ src/comments/constants.ts
@@ -1,0 +1,3 @@
+export const GRAVATAR_CONTAINER = '.gravatar-enhanced-profile';
+export const COMMENT_EMAIL_WRAPPER = '.comment-form-email';
+export const COMMENT_EMAIL_FIELD = '#email';
